requirements.txt
scrapling[fetchers]
google-adk[a2a]
uvicorn
python-dotenv
google-generativeai
google-genai
fastapi
httpx
websockets
